requirements.txt
asgiref==3.9.1
bcrypt==4.3.0
cffi==1.17.1
cryptography==45.0.6
Django==5.2.5
django-mysql==4.17.0
djangorestframework==3.16.1
pycparser==2.22
PyMySQL==1.1.1
sqlparse==0.5.3
tzdata==2025.2
